// src/infra/api/model/avaliacaoModel.ts
/**
 * Generated by orval v8.4.1 🍺
 * Do not edit manually.
 * Minha API
 * Documentação oficial da API
 * OpenAPI spec version: 1.0.0
 */

export interface AvaliacaoModel {
  id: number;
  /** @nullable */
  estabelecimentoId: number | null;
  userId: number;
  rating: number;
  /** @nullable */
  comentario: string | null;
  /** @nullable */
  approved: boolean | null;
  /** @nullable */
  createdAt: string | null;
  /** @nullable */
  updatedAt: string | null;
  /** @nullable */
  deletedAt: string | null;
}